refactor: move signAndBroadcast to lib
